telegram/requirements.txt
aiogram>=3.0.0
httpx[http2]
python-dotenv
